complete header styling
