--- src/components/Header/Brand.jsx
@@ -1,12 +1,0 @@
-import classes from "./Brand.module.css";
-
-function Brand() {
-  return (
-    <section className={classes.brand}>
-      <h1>gruppo</h1>
-      <p>categorize your resources</p>
-    </section>
-  );
-}
-
-export default Brand;
