--- src/components/Header/Header.module.css
@@ -1,6 +1,0 @@
-.header {
-  grid-area: a;
-  background: rgb(82, 94, 117);
-  color: whitesmoke;
-  padding: 1rem;
-}
