--- src/components/Header/Tags.jsx
@@ -1,5 +1,0 @@
-function Tags() {
-  return;
-}
-
-export default Tags;
