--- src/components/Header/User.jsx
@@ -1,5 +1,0 @@
-function User() {
-  return;
-}
-
-export default User;
