+++ src/features/Content/Content.jsx
@@ -1,0 +1,11 @@
+import styles from "./Content.module.css";
+
+function Content() {
+  return (
+    <section className={styles.content}>
+      <h2>My resources</h2>
+    </section>
+  );
+}
+
+export default Content;
