+++ src/features/Content/Content.module.css
@@ -1,0 +1,11 @@
+.content {
+  grid-area: b;
+  padding: 1rem;
+}
+
+/* For Desktop View */
+@media screen and (min-width: 1024px) {
+  .content {
+    padding: 2rem;
+  }
+}
