+++ src/features/Header/User.jsx
@@ -1,0 +1,24 @@
+import userStyles from "./User.module.css";
+import buttonStyles from "../../components/Button.module.css";
+
+const styles = { ...userStyles, ...buttonStyles };
+
+function User() {
+  return (
+    <div className={styles.user}>
+      {/*todo use css modules compose to reduce*/}
+      <button
+        className={`${styles.avatar} ${styles.btn} ${styles.md} ${styles["secondary-outline"]} ${styles.rounded}`}
+      >
+        Tim Apple
+      </button>
+      <ul className={styles.actions}>
+        <li>
+          <a href="">Logout</a>
+        </li>
+      </ul>
+    </div>
+  );
+}
+
+export default User;
